Fixed some conflicting types issues, and apparently lame deprecated some of its interface, so had to do some replacements

+++ src/in_flac.c
@@ -40,7 +40,7 @@ typedef struct {
 } flac_in_t;
 
 /* -- static prototypes -- */
-static int ices_flac_readpcm (input_stream_t* self, size_t len,
+static ssize_t ices_flac_readpcm (input_stream_t* self, size_t len,
                               int16_t* left, int16_t* right);
 static int ices_flac_close (input_stream_t* self);
